# Conflicts:
#	composer.lock
#	src/Controller/TeamController.php
#	src/Controller/TournamentController.php
#	src/Entity/Team.php
#	templates/tournament/infos.html.twig

+++ src/Controller/TournamentController.php
@@ -11,6 +11,7 @@ use Doctrine\ORM\EntityManagerInterface;
 use Symfony\Bridge\Doctrine\Form\Type\EntityType;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\Form\Extension\Core\Type\SubmitType;
+use Symfony\Bundle\SecurityBundle\Security;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
@@ -34,6 +35,7 @@ class TournamentController extends AbstractController
     }
 
     #[Route('/nouveau', name: 'tournois_nouveau', methods: ['GET', 'POST'])]
+    #[IsGranted('ROLE_USER')]
     public function nouveau(
         Request $request,
         EntityManagerInterface $entityManager,
@@ -58,6 +60,11 @@ class TournamentController extends AbstractController
         //dd($result);
 
         $tournament = new Tournament();
+        //Récupération du statut pour éviter la modification dans le formulaire
+        //$status = $tournament->getStatus();
+
+        //Récupération de la date de création pour éviter la modification dans le formulaire
+        //$dateCreation = $tournament->getDateCreation();
 
         $form = $this->createForm(TournamentType::class, $tournament);
         $form->handleRequest($request);
@@ -66,7 +73,9 @@ class TournamentController extends AbstractController
         $tournament->setUser($user);
 
         if ($form->isSubmitted() && $form->isValid()) {
+            $user->setRoles(["ROLE_ORGANIZER"]);
             $entityManager->persist($tournament);
+            $entityManager->persist($user);
             $entityManager->flush();
 
             return $this->redirectToRoute('tournois_home', [], Response::HTTP_SEE_OTHER);
@@ -81,10 +90,19 @@ class TournamentController extends AbstractController
     }
 
     #[Route('/{id}', name: 'tournois_infos', methods: ['GET'])]
-    public function infos(Tournament $tournament): Response
+
+    #[IsGranted('ROLE_USER')]
+    public function infos(Tournament $tournament, TeamRepository $teamRepository): Response
+
     {
         $teams = $tournament->getTeam();
         //dd($teams->count());
+
+        $teams = $teamRepository->findAll();
+
+        // Mélange des équipes
+        // TODO : A mettre en base de donnée pour stocker 1 fois (nouvelle entité ?)
+        shuffle($teams);
 
         return $this->render('tournament/infos.html.twig', [
             'tournament' => $tournament,
@@ -157,35 +175,63 @@ class TournamentController extends AbstractController
         name: 'tournois_modifier',
         methods: ['GET', 'POST']
     )]
-    public function modifier(Request $request, Tournament $tournament, EntityManagerInterface $entityManager): Response
+    #[IsGranted('ROLE_ORGANIZER')]
+    public function modifier(Request $request, Tournament $tournament, EntityManagerInterface $entityManager, Security $security): Response
     {
         $dateCreation = $tournament->getDateCreation();
         $status = $tournament->getStatus();
 
-        $form = $this->createForm(TournamentType::class, $tournament);
-        $form->handleRequest($request);
+        // Récupérer l'utilisateur connecté
+        $user = $security->getUser();
 
-        if ($form->isSubmitted() && $form->isValid()) {
-            $entityManager->flush();
+        // Vérifier si l'utilisateur est connecté et son ID correspond à l'ID du tournoi
+        if ($user && $user->getId() === $tournament->getUser()->getId()){
+            // Vérifier si l'utilisateur a le rôle nécessaire
+            if ($this->isGranted('ROLE_ORGANIZER', $user) || $this->isGranted('ROLE_ADMIN', $user)) {
+                $form = $this->createForm(TournamentType::class, $tournament);
+                $form->handleRequest($request);
 
-            return $this->redirectToRoute('tournois_home', [], Response::HTTP_SEE_OTHER);
+                if ($form->isSubmitted() && $form->isValid()) {
+                    $entityManager->flush();
+
+                    return $this->redirectToRoute('tournois_home', [], Response::HTTP_SEE_OTHER);
+                }
+
+                return $this->render('tournament/modifier.html.twig', [
+                    'tournament' => $tournament,
+                    'form' => $form,
+                    'dateCreation' => $dateCreation,
+                    'status' => $status
+                ]);
+            } else {
+                // L'utilisateur n'a pas les droits nécessaires
+                /*throw $this->createAccessDeniedException();*/
+                return $this->redirectToRoute('error_403', [], Response::HTTP_SEE_OTHER);
+            }
+        } else {
+            // L'utilisateur n'est pas autorisé à accéder à cette page
+            return $this->redirectToRoute('error_403', [], Response::HTTP_SEE_OTHER);
         }
-
-        return $this->render('tournament/modifier.html.twig', [
-            'tournament' => $tournament,
-            'form' => $form,
-            'dateCreation' => $dateCreation,
-            'status' => $status
-        ]);
     }
 
     #[Route('/{id}', name: 'tournois_supprimer', methods: ['POST'])]
-    public function supprimer(Request $request, Tournament $tournament, EntityManagerInterface $entityManager): Response
+    #[IsGranted('ROLE_ORGANIZER')]
+    public function supprimer(Request $request, Tournament $tournament, EntityManagerInterface $entityManager, Security $security): Response
     {
-        if ($this->isCsrfTokenValid('delete'.$tournament->getId(), $request->request->get('_token'))) {
-            $entityManager->remove($tournament);
-            $entityManager->flush();
-        }
+        // Récupérer l'utilisateur connecté
+        $user = $security->getUser();
+
+        // Vérifier si l'utilisateur est connecté et son ID correspond à l'ID du tournoi
+        if ($user && $user->getId() === $tournament->getUser()->getId()) {
+            // Vérifier si l'utilisateur a le rôle nécessaire
+            if ($this->isGranted('ROLE_ORGANIZER', $user)) {
+                if ($this->isCsrfTokenValid('delete'.$tournament->getId(), $request->request->get('_token'))) {
+                    $entityManager->remove($tournament);
+                    $entityManager->flush();
+                } else {
+                    // Le jeton CSRF n'est pas valide, vous pouvez gérer cela en lançant une exception appropriée
+                    throw new \Exception('Jeton CSRF invalide.');
+                }
 
         return $this->redirectToRoute('tournois_home', [], Response::HTTP_SEE_OTHER);
     }
